requirements.txt
discord.py[voice]==2.3.2
discord-ext-voice-recv
PyNaCl==1.5.0
openai>=2.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp>=3.8.0

